feat: :lock: change path bug

+++ server/server.js
@@ -17,6 +17,14 @@ app.get("/list", (req, res) => {
 
 app.get("/download/:filename", (req, res) => {
   let file_name = req.params.filename;
+  if (
+    file_name.includes("..") ||
+    file_name.includes("/") ||
+    file_name.includes("\\") ||
+    file_name.includes("%")
+  ) {
+    res.status(400).json({ error: "Invalid file name" });
+  }
   let file_path = filem.getFilePath(file_name);
   if (!file_path) return res.status(404).send("File not found");
   return res.download(file_path);
@@ -24,6 +32,9 @@ app.get("/download/:filename", (req, res) => {
 
 app.post("/delete", (req, res) => {
   let { file_name } = req.body;
+  if (!filem.security_check(file_name)) {
+    return res.status(400).json({ error: "Invalid file name" });
+  }
   if (filem.delete(file_name)) {
     return res.json({ success: true });
   }
@@ -33,6 +44,9 @@ app.post("/delete", (req, res) => {
 app.post("/upload", upload.array("files"), (req, res) => {
   if (!req.files) return res.status(400).send("No files were uploaded.");
   req.files.forEach((file) => {
+    if (!filem.security_check(file.originalname)) {
+      return res.status(400).json({ error: "Invalid file name" })
+    }
     if (!filem.create(file.originalname, file.buffer)) {
       return res.json({ success: false, message: "Error creating file" });
     }
